refactor:抽象Builder和Config抽象类
